Pull model lookup into shared function
Fix a bug where use-model stored incorrect model id
