+++ cmd/modelctl/commands/show-model.go
@@ -98,49 +98,13 @@ func (cmd *showModelCommand) showCurrentModel() error {
 	return cmd.showModel(currentModel)
 }
 
-func (cmd *showModelCommand) showModel(modelName string) error {
-	allModelManifests, err := models.LoadManifests(cmd.ModelsDir)
+func (cmd *showModelCommand) showModel(modelNameOrID string) error {
+	modelManifest, err := common.GetModelByNameOrId(cmd.Context, modelNameOrID)
 	if err != nil {
-		return fmt.Errorf("loading models: %v", err)
+		return err
 	}
 
-	activeEngine, err := cmd.Cache.GetActiveEngine()
-	if err != nil {
-		return fmt.Errorf("%s: %w", common.LookingUpActiveEngine, err)
-	}
-	if activeEngine == "" {
-		return common.ErrNoActiveEngine
-	}
-
-	engineManifest, err := engines.LoadManifest(cmd.EnginesDir, activeEngine)
-	if err != nil {
-		return fmt.Errorf("loading manifest: %v", err)
-	}
-
-	// Consider the active engine's models first
-	var manifest *models.Manifest
-	for _, modelManifest := range allModelManifests {
-		if slices.Contains(engineManifest.Model.Options, modelManifest.ID) &&
-			(modelManifest.Name == modelName || modelManifest.ID == modelName) {
-			manifest = &modelManifest
-			break
-		}
-	}
-
-	// If the provided name is not one of the active engine's models, warn the user it is not compatible
-	if manifest == nil {
-		for _, modelManifest := range allModelManifests {
-			if modelManifest.Name == modelName || modelManifest.ID == modelName {
-				return fmt.Errorf("model %q is not compatible with active engine", modelName)
-			}
-		}
-	}
-
-	if manifest == nil {
-		return fmt.Errorf("model %q does not exist", modelName)
-	}
-
-	err = cmd.printModelManifest(manifest)
+	err = cmd.printModelManifest(modelManifest)
 	if err != nil {
 		return fmt.Errorf("printing model manifest: %v", err)
 	}
