+++ test_data/models/26b-q4-k-m-gguf/model.yaml
@@ -1,4 +1,4 @@
-id: 26b-a4b-it-q4-k-m-gguf
+id: 26b-q4-k-m-gguf
 name: 26b
 
 description: Test model description
